% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.Beta.EB.R
\name{binom.Beta.EB}
\alias{binom.Beta.EB}
\title{Empirical Bayes Beta Prior for Binomial Data}
\usage{
binom.Beta.EB(x, n, verbose = FALSE)
}
\arguments{
\item{x}{historical events}

\item{n}{historical trials}

\item{verbose}{Print messages}
}
\value{
A function of the probability parmater p
}
\description{
Empirical Bayes Beta Prior for Binomial Data
}

