% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.Bias.R
\name{binom.Bias.FB}
\alias{binom.Bias.FB}
\title{Full Bayes Bias Prior for Binomial Data}
\usage{
binom.Bias.FB(x, n, verbose = FALSE, length = 30, dprior, mc.cores = 1)
}
\arguments{
\item{x}{historical events}

\item{n}{historical trials}

\item{tau.prior}{optional prior on bias parameter}
}
\value{
A function of the probability parmater p
}
\description{
Full Bayes Bias Prior for Binomial Data
}
\examples{

}

