% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.MAP.FB.R
\name{binom.MAP.FB}
\alias{binom.MAP.FB}
\title{Full Bayesian Meta-Analytic Prior for Binomial Data}
\usage{
binom.MAP.FB(x, n, tau.prior, verbose = FALSE)
}
\arguments{
\item{x}{number of historical successes}

\item{n}{number historical patients}

\item{tau.prior}{optional prior on heterogeneity parameter}

\item{verbose}{Print messages}
}
\value{
A function of the probability parmater p
}
\description{
Full Bayesian Meta-Analytic Prior for Binomial Data
}
\examples{
\donttest{
xh <- c(30,40,50)
nh <- c(90,95,110)

# Fit a MAP prior
map <- binom.MAP.FB(xh,nh, X=34,N=75, tau.prior = list(prior= "logtnormal", param=c(0,0.5)))

# Approximate with beta densities and add a 20\% robust component
mix.map <- conj.approx(map, "beta", robust=0.2) 
 }
}
