% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.EB.Beta.R
\name{binom.PP.EB.Beta}
\alias{binom.PP.EB.Beta}
\title{Power Prior for Binomial Data Using Empirical Bayes on Beta Prior on Weights}
\usage{
binom.PP.EB.Beta(x, n, X, N, verbose = FALSE)
}
\arguments{
\item{x}{number of historical successes}

\item{n}{number historical patients}

\item{X}{number of new successes}

\item{N}{number of new patients}

\item{verbose}{Print messages}
}
\value{
A function of the probability parmater p
}
\description{
Power Prior for Binomial Data Using Empirical Bayes on Beta Prior on Weights
}

