% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.EB.Sep.R
\name{binom.PP.EB.Sep}
\alias{binom.PP.EB.Sep}
\title{Empirical Bayes Power Prior for Binomial Data Esimated Separately}
\usage{
binom.PP.EB.Sep(x, n, X, N, verbose = FALSE)
}
\arguments{
\item{x}{number of historical successes}

\item{n}{number historical patients}

\item{X}{number of new successes}

\item{N}{number of new patients}

\item{verbose}{Print messages}
}
\value{
A function of the probability parmater p
}
\description{
Empirical Bayes Power Prior for Binomial Data Esimated Separately
}

