% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.FB.INLA.R
\name{binom.PP.FB.INLA}
\alias{binom.PP.FB.INLA}
\title{Title}
\usage{
binom.PP.FB.INLA(x, n, X, N, verbose = FALSE, length = 30, d.prior.a = 1,
  d.prior.b = 1, mc.cores = 1)
}
\arguments{
\item{d.prior.a}{shape1 parameter for beta prior on weight}

\item{d.prior.b}{shape2 parameter for beta prior on weight}

\item{focus}{}
}
\description{
Title
}

