% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.FB.MC.BE.R
\name{binom.PP.FB.MC.BE}
\alias{binom.PP.FB.MC.BE}
\title{Full Bayes Power Prior with Beta Prior on Weights}
\usage{
binom.PP.FB.MC.BE(x, n, verbose = FALSE, length = 30, d.prior.a = 1,
  d.prior.b = 1, p.prior.a = 1, p.prior.b = 1, mc.cores = 1,
  samples = 10000, focus)
}
\arguments{
\item{x}{number of historical successes}

\item{n}{number historical patients}

\item{verbose}{Print messages}

\item{length}{Number of points to evaluate density at}

\item{d.prior.a}{shape1 parameter for beta prior on weights}

\item{d.prior.b}{shape2 parameter for beta prior on weights}

\item{p.prior.a}{shape1 parameter for beta prior on probability}

\item{p.prior.b}{shape2 parameter for beta prior on probability}

\item{mc.cores}{Number of cores for parallel}

\item{samples}{Number of Monte Carlo samples}

\item{focus}{List of triples specifying regions to focus on, eg peaks. Specified as list(c(lower, upper, length))}
}
\value{
A density function
}
\description{
Full Bayes Power Prior with Beta Prior on Weights
}

