% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.FB.R
\name{binom.PP.FB}
\alias{binom.PP.FB}
\title{Full Bayes Power Prior for Binomial Data}
\usage{
binom.PP.FB(x, n, verbose = FALSE, length = 30, dprior, mc.cores = 1,
  p.prior.a = 1, p.prior.b = 1)
}
\arguments{
\item{x}{historical events}

\item{n}{historical trials}

\item{tau.prior}{optional prior on heterogeneity parameter}
}
\value{
A function of the probability parmater p
}
\description{
Full Bayes Power Prior for Binomial Data
}
\examples{

}

