% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.prior.R
\name{binom.prior}
\alias{binom.prior}
\title{Prior distribution for binomial data}
\usage{
binom.prior(type = c("MAP.FB", "PP.FB", "PP.EB", "Beta.EB", "PP.EB.Sep",
  "PP.Fix"), x, n, verbose = FALSE, ...)
}
\arguments{
\item{type}{Which prior to provide}

\item{x}{Vector of historical events}

\item{n}{Vector of historical sample sizes}

\item{verbose}{Print messages}

\item{...}{Parameters to provide to prior}
}
\value{
A density function
}
\description{
Prior distribution for binomial data
}
\examples{
\donttest{
xh <- c(30,40,50)
nh <- c(90,95,110)
eb <- binom.prior("PP.EB", xh, nh, X=34,N=75)
fb <- binom.prior("PP.FB", xh, nh)

p <- seq(0,1,len=100)
plot(p,fb(p), ty='l', ylim=c(0,15))
lines(p, eb(p,X=0), col=2)
}

}
