% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bool.mat.plot.R
\name{bool.mat.plot}
\alias{bool.mat.plot}
\title{Plot a boolean matrix in two colours}
\usage{
bool.mat.plot(mat, col.pair = c("darkred", "lightblue"), ...)
}
\arguments{
\item{mat}{Matrix with TRUE or FALSE entries}

\item{col.pair}{Vector of colours. First entry is used for FALSE and second for TRUE}

\item{...}{extra parameters to plot}
}
\value{
Nothing
}
\description{
Plot a boolean matrix in two colours
}

