% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.mean.R
\name{calc.MSE.mean}
\alias{calc.MSE.mean}
\title{Calculate mean squared error based on the mean of the posterior}
\usage{
calc.MSE.mean(prior, prob.range = c(0.5, 1), length = 20, n.binom = 30,
  mc.cores = 1, posterior)
}
\arguments{
\item{prior}{Prior to calculate posterior. Specify posterior instead if available.}

\item{prob.range}{Range of values to calculate MSE over}

\item{length}{Number of values to calculate MSE for}

\item{n.binom}{Number of patients in new trial}

\item{mc.cores}{Number of cores for parallel}

\item{posterior}{Posterior density}
}
\value{
A vector of error values
}
\description{
Calculate mean squared error based on the mean of the posterior
}

