% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.mode.R
\name{calc.MSE.mode}
\alias{calc.MSE.mode}
\title{Calculate mean squared error based on the mode of the posterior}
\usage{
calc.MSE.mode(
  prior,
  prob.range = c(0.5, 1),
  length = 20,
  n.binom = 30,
  mc.cores = 1,
  posterior
)
}
\arguments{
\item{prior}{Prior to calculate posterior. Specify posterior instead if available.}

\item{prob.range}{Range of values to calculate MSE over}

\item{length}{Number of values to calculate MSE for}

\item{n.binom}{Number of patients in new trial}

\item{mc.cores}{Number of cores for parallel}

\item{posterior}{Posterior density}
}
\value{
A vector of error values
}
\description{
Calculate mean squared error based on the mode of the posterior
}
\examples{
\donttest{
xh <- c(30,40,50)
nh <- c(90,95,110)
fix <- binom.PP.FIX(x=xh, n=nh, d=c(.2,.5,.7), mix=TRUE) # set different weights
calc.MSE.mode(prior = fix, prob.range=c(0,1), n.binom=100)
}

}
