% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.posterior.R
\name{calc.posterior}
\alias{calc.posterior}
\title{Calculate a posterior density based on a binomial likelihood}
\usage{
calc.posterior(prior, X, N)
}
\arguments{
\item{prior}{Prior}

\item{X}{Number of successes}

\item{N}{Number of patients}
}
\value{
The posterior density function
}
\description{
Calculate a posterior density based on a binomial likelihood
}

