% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.3.robust.R
\name{conj.approx.robust}
\alias{conj.approx.robust}
\title{Approximate distribution with mixture model}
\usage{
conj.approx.robust(distr, type = c("beta", "normal"),
  min.degree = max.degree, max.degree = 3, return.value = 0.2,
  range = c(0, 1), starts, length.fit = 100, do.plot = FALSE)
}
\arguments{
\item{length.fit}{}
}
\description{
Approximate distribution with mixture model
}

