% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.2.R
\name{conj.approx2}
\alias{conj.approx2}
\title{Approximate distribution with mixture model}
\usage{
conj.approx2(distr, type = c("beta", "normal"), degree = 3, range = c(0,
  1), starts, length.fit = 100, do.plot = FALSE)
}
\arguments{
\item{length.fit}{}
}
\description{
Approximate distribution with mixture model
}

