% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.rw.R
\name{create.mixture.prior}
\alias{create.mixture.prior}
\title{Create mixture model}
\usage{
create.mixture.prior(type, pars, weights = rep(1/nrow(pars), nrow(pars)))
}
\arguments{
\item{type}{Type of mixture, "normal" or "beta"}

\item{pars}{Matrix of parameter values}

\item{weights}{Mixture weights}
}
\value{
A \code{mixture.prior} object
}
\description{
Create mixture model
}

