% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.continuous.R
\name{dbinomc}
\alias{dbinomc}
\title{Extension of the binomial distribution to allow real values for size and x}
\usage{
dbinomc(x, size, prob)
}
\arguments{
\item{prob}{}
}
\description{
Extension of the binomial distribution to allow real values for size and x
}

