% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{ess.fun.list}
\alias{ess.fun.list}
\title{Calculate the effective sample size of mixture model}
\usage{
ess.fun.list(fun.list)
}
\arguments{
\item{fun.list}{}
}
\description{
Calculate the effective sample size of mixture model
}

