% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.rw.R
\name{ess.mixture.prior}
\alias{ess.mixture.prior}
\title{Calculate the equivalent sample size of mixture model}
\usage{
ess.mixture.prior(mixture.prior)
}
\arguments{
\item{mixture.prior}{Mixture prior object}
}
\value{
Sample size
}
\description{
Calculate the equivalent sample size of mixture model
}
\examples{
\donttest{
xh <- c(30,40,50)
nh <- c(90,95,110)
fb <- binom.PP.FB.COR(x=xh, n=nh, mixture.size=500, mix=TRUE)
fb2 <- binom.PP.FB.COR(x=xh, n=nh, mixture.size=500, mix=TRUE,d.prior.cor=0.9) 
ess.mixture.prior(fb2)
ess.mixture.prior(fb)
}
}
