% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.morita.R
\name{ess.morita.mixture.prior}
\alias{ess.morita.mixture.prior}
\title{Prior effective sample size}
\usage{
ess.morita.mixture.prior(mixture, c = 100, LEN = 1000)
}
\arguments{
\item{mixture}{A mixture.prior object}

\item{c}{factor such that Beta( mu/c , (1-mu)/c ) is non0informative, where mu is mode of prior}

\item{LEN}{Number of points to evaluate at. Fewer is faster}
}
\value{
Estimate of the effective sample size.
}
\description{
Estimate the prior effecive sample size for Beta-mixture
 priors based on Morita, Thall, Mueller (2008, Biometrics)
 evaluated at the prior mode.
}
