% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.rw.R
\name{eval.mixture.prior}
\alias{eval.mixture.prior}
\title{Get density of mixture}
\usage{
eval.mixture.prior(x, mixture.prior, subset)
}
\arguments{
\item{x}{Vector of values to evaluate density at}

\item{mixture.prior}{Mixture prior}

\item{subset}{Vector of values specifying which mixture components should be included, eg c(1,2)}
}
\value{
Densities at of prior at x
}
\description{
Get density of mixture
}

