% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eb_powers.R
\name{get_eb_powers}
\alias{get_eb_powers}
\title{Get Power Estimates from EB functions}
\usage{
get_eb_powers(pp_fun)
}
\arguments{
\item{pp_fun}{A power prior function from [binom.PP.EB()] with `mix=FALSE`)}
}
\value{
A `data.frame` containing the power prior weights
}
\description{
Get Power Estimates from EB functions
}
\examples{
pp_eb <- binom.PP.EB(x = c(24,35), n = c(50,50), X = 0:50, N=50, mix = FALSE)
get_eb_powers(pp_eb)
}
