% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{install.inla}
\alias{install.inla}
\title{Install R-INLA from http://www.r-inla.org/download}
\usage{
install.inla(testing = FALSE)
}
\arguments{
\item{testing}{install the testing version. Default (FALSE) installs the stable version.}
}
\value{
TRUE if INLA can be successfully loaded after installation.
}
\description{
Install R-INLA from http://www.r-inla.org/download
}
