% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{is.mixture.prior}
\alias{is.mixture.prior}
\title{Test if object is a mixture.prior}
\usage{
is.mixture.prior(x)
}
\arguments{
\item{x}{Object}
}
\value{
TRUE if \code{x} is a \code{mixture.prior}, else FALSE
}
\description{
Test if object is a mixture.prior
}

