% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.rw.R
\name{make.robust}
\alias{make.robust}
\title{Add a vague componenet to mixture model to make robust}
\usage{
make.robust(object, weight)
}
\arguments{
\item{object}{mixture.prior to update}

\item{weight}{weight of vague component in updated model}
}
\value{
A \code{mixture.prior} object with updated weights and parameters
}
\description{
Add a vague componenet to mixture model to make robust
}
