% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{mean.fun.list}
\alias{mean.fun.list}
\title{Calculate the mean of mixture model}
\usage{
\method{mean}{fun.list}(fun.list)
}
\arguments{
\item{fun.list}{}
}
\description{
Calculate the mean of mixture model
}

