% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{mean.mixture.prior}
\alias{mean.mixture.prior}
\title{Calculate the mean of mixture model}
\usage{
\method{mean}{mixture.prior}(x, ...)
}
\arguments{
\item{x}{Mixture prior object}

\item{...}{Ignored}
}
\value{
Mean of prior
}
\description{
Calculate the mean of mixture model
}

