% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.plot.R
\name{mix.plot}
\alias{mix.plot}
\title{Plot the mixture model}
\usage{
mix.plot(pars, weights, dist, split.f)
}
\arguments{
\item{pars}{Parameters of each distribution component}

\item{weights}{Vector of weights for each component}

\item{dist}{Distribution}
}
\description{
Plot the mixture model
}

