% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.PP.EB.R
\name{normal.PP.EB}
\alias{normal.PP.EB}
\title{Empirical Bayes power prior for normal likelihoods}
\usage{
normal.PP.EB(x, sd, X, SD)
}
\arguments{
\item{x}{vector of historical means}

\item{sd}{vector of historical standard deviations}

\item{X}{new study mean}

\item{SD}{new study standard deviation}
}
\value{
A density function of the mean parameter
}
\description{
Empirical Bayes power prior for normal likelihoods
}
