% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.PP.FB.Cor.R
\name{normal.PP.FB.COR}
\alias{normal.PP.FB.COR}
\title{Full Bayes Power Prior with Correlated Weight Parameters using Rao-Blackwell}
\usage{
normal.PP.FB.COR(
  x,
  sd,
  verbose = FALSE,
  mixture.size = 1000,
  d.prior.cor = 0,
  mc.cores = 1
)
}
\arguments{
\item{x}{vector of means of historical studies}

\item{sd}{vector of standard deviations of historical studies}

\item{verbose}{Print messages}

\item{mixture.size}{Number of components in the mixture. If d.prior.cor=0, will be rounded to nearest value a^length(x) to make the grid evenly spaced.}

\item{d.prior.cor}{Correlation parameter for transformed multivariate normal prior on weights}

\item{mc.cores}{Number of cores for parallel}
}
\value{
A density function
}
\description{
Full Bayes Power Prior with Correlated Weight Parameters using Rao-Blackwell
}
