% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.PP.FB.MC.R
\name{normal.PP.FB.MC}
\alias{normal.PP.FB.MC}
\title{Normal power prior with full Bayes by Monte Carlo}
\usage{
normal.PP.FB.MC(x, sd, verbose = FALSE, length = 30, d.prior.a = 1,
  d.prior.b = 1, mc.cores = 1, samples = 10000, focus)
}
\arguments{
\item{x}{numeric vector of results}

\item{sd}{standard deviations}

\item{verbose}{Print messages}

\item{length}{Number of points to evaluate density at}

\item{d.prior.a}{shape1 parameter for beta prior on weights}

\item{d.prior.b}{shape2 parameter for beta prior on weights}

\item{mc.cores}{Number of cores for parallel}

\item{samples}{Number of Monte Carlo samples}

\item{focus}{List of triples specifying regions to focus on, eg peaks. Specified as list(c(lower, upper, length))}
}
\value{
Density function
}
\description{
Normal power prior with full Bayes by Monte Carlo
}

