% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.rw.R
\name{p.mixture.prior}
\alias{p.mixture.prior}
\title{Probabilty function}
\usage{
p.mixture.prior(q, mixture.prior)
}
\arguments{
\item{q}{Quantile}

\item{mixture.prior}{Mixture prior object}
}
\value{
Vectore of probabilities
}
\description{
Probabilty function
}

