% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.rw.R
\name{plot.mixture.prior}
\alias{plot.mixture.prior}
\title{Plot mixture model}
\usage{
\method{plot}{mixture.prior}(x, ..., at = seq(0, 1, 0.005), stack = FALSE, add = FALSE)
}
\arguments{
\item{x}{a mixture prior object to plot}

\item{...}{Additional arguments to \code{\link{lines}}}

\item{at}{Vector of values to evaluate density at}

\item{stack}{Plot the mixture components as stacked regions}

\item{add}{lines to an existing plot}
}
\description{
Plot mixture model
}
\examples{
\donttest{
xh <- c(30,40,50)
nh <- c(90,95,110)

#fit a full Bayes power prior with 500 samples
fb <- binom.PP.FB(x=xh, n=nh, mixture.size=500, mix=TRUE) 

#fit a full Bayes power prior with 500 samples and correlation 0.9
fb2 <- binom.PP.FB(x=xh, n=nh, mixture.size=500, mix=TRUE,d.prior.cor=0.9) 
plot.mixture.prior(fb2, xlab="p",ylab="Density")
plot.mixture.prior(fb, lines.only=TRUE, col=2)
}
}
