% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{posterior.fun.list}
\alias{posterior.fun.list}
\title{Calculate posterior mixture model}
\usage{
posterior.fun.list(xs, ns, fun.list)
}
\arguments{
\item{fun.list}{}
}
\description{
Calculate posterior mixture model
}

