% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{print.fun.list}
\alias{print.fun.list}
\title{Print mixture model}
\usage{
\method{print}{fun.list}(fun.list)
}
\arguments{
\item{fun.list}{The mixture model to print}
}
\description{
Print mixture model
}

