% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{print.mixture.prior}
\alias{print.mixture.prior}
\title{Print mixture model}
\usage{
\method{print}{mixture.prior}(x, ...)
}
\arguments{
\item{x}{Mixture prior object}

\item{...}{Ignored}
}
\description{
Print mixture model
}

