% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.matrix.R
\name{prob.control.larger}
\alias{prob.control.larger}
\alias{prob.control.smaller}
\title{Compare the the control parameter and the treatment parameter}
\usage{
prob.control.larger(xt, nt, posterior, xc, nc, prior)

prob.control.smaller(xt, nt, posterior, xc, nc, prior)
}
\arguments{
\item{xt}{number of events in the treatment arm}

\item{nt}{number of of patients in the treatment arm}

\item{posterior}{Posterior mixture.prior object}

\item{xc}{number of events in the control arm}

\item{nc}{number of patients in the control}

\item{prior}{mixture.prior object to calculate posterior using nc and xc}
}
\value{
The probability the control parameter is smaller/larger than the treatment parameter.
\code{prob.control.smaller} is \code{1-prob.control.larger} and is provided for convenience.
}
\description{
Compare the the control parameter and the treatment parameter
}
