% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.rw.R
\name{q.mixture.prior}
\alias{q.mixture.prior}
\title{Quantile function}
\usage{
q.mixture.prior(p, mixture.prior)
}
\arguments{
\item{p}{Probability}

\item{mixture.prior}{Mixture prior object}
}
\value{
Vector of quantiles
}
\description{
Quantile function
}

