% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{sample.fun.list}
\alias{sample.fun.list}
\title{Sample from the mixture distribution}
\usage{
sample.fun.list(n, fun.list)
}
\arguments{
\item{n}{number of samples}

\item{fun.list}{}
}
\value{
A vector of n observations
}
\description{
Sample from the mixture distribution
}

