% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{sample.mixture.prior}
\alias{sample.mixture.prior}
\title{Sample from the mixture distribution}
\usage{
sample.mixture.prior(n, mixture.prior)
}
\arguments{
\item{n}{Number of samples}

\item{mixture.prior}{Mixture prior object}
}
\value{
A vector of n samples from \code{mixture.prior}
}
\description{
Sample from the mixture distribution
}

