% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.matrix.R
\name{sig.matrix}
\alias{sig.matrix}
\title{Title}
\usage{
sig.matrix(n.control, n.treatment, level = 0.975, prior,
  treat.beta.prior.par = c(1, 1), mc.cores = 1, check.xt, check.xs,
  debug = FALSE)
}
\arguments{
\item{n.control}{Number of subjects in the control group}

\item{level}{Significance level to test (default 0.975)}

\item{prior}{Prior on binomial probability parameter of the control group}

\item{treat.beta.prior.par}{Paramters for beta prior on treatment group probability parameter}

\item{mc.cores}{Number of cores to use for mclapply}
}
\value{
A matrix of TRUE/FALSE values of size n.control+1 x n.treatment+1
}
\description{
Title
}

