% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{update.fun.list}
\alias{update.fun.list}
\title{Update parameters or weights in mixture model}
\usage{
update.fun.list(fun.list, pars, weights)
}
\arguments{
\item{weights}{}
}
\description{
Update parameters or weights in mixture model
}

