% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{var.fun.list}
\alias{var.fun.list}
\title{Calculate the variance of mixture model}
\usage{
var.fun.list(fun.list)
}
\arguments{
\item{fun.list}{}
}
\description{
Calculate the variance of mixture model
}

