% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.rw.R
\name{var.mixture.prior}
\alias{var.mixture.prior}
\title{Calculate the variance of mixture model}
\usage{
var.mixture.prior(mixture.prior)
}
\arguments{
\item{mixture.prior}{Mixture prior object}
}
\value{
Variance of prior
}
\description{
Calculate the variance of mixture model
}

