% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VP.R
\name{VP}
\alias{VP}
\title{Print messages if verbose=TRUE in calling function.}
\usage{
VP(message)
}
\arguments{
\item{message}{Message to print}
}
\description{
Print messages if verbose=TRUE in calling function.
}

