% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.Bias.R
\name{binom.Bias.FB}
\alias{binom.Bias.FB}
\title{Full Bayes Bias Prior for Binomial Data}
\usage{
binom.Bias.FB(x, n, verbose = FALSE, mc.cores = 1)
}
\arguments{
\item{x}{number of historical successes}

\item{n}{number historical patients}

\item{verbose}{Print messages}

\item{mc.cores}{number of cores for parallel}
}
\value{
A function of the probability parmater p
}
\description{
Full Bayes Bias Prior for Binomial Data
}

