% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.MAP.EB.R
\name{binom.MAP.EB}
\alias{binom.MAP.EB}
\title{Empirical Bayes Meta-Analytic Prior for Binomial Data}
\usage{
binom.MAP.EB(x, n, X, N, verbose = FALSE, upper, VAR)
}
\arguments{
\item{x}{historical events}

\item{n}{historical trials}

\item{tau.prior}{optional prior on heterogeneity parameter}
}
\value{
A function of the probability parmater p
}
\description{
Empirical Bayes Meta-Analytic Prior for Binomial Data
}
\examples{

}

