% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.MAP.EB.R
\name{binom.MAP.EB}
\alias{binom.MAP.EB}
\title{Empirical Bayes Meta-Analytic Prior for Binomial Data}
\usage{
binom.MAP.EB(x, n, X, N, verbose = FALSE, upper = 4, mc.cores)
}
\arguments{
\item{x}{number of historical successes}

\item{n}{number historical patients}

\item{X}{vector of new successes to computer prior for}

\item{N}{number of new patients}

\item{verbose}{Print messages}

\item{upper}{Upper limit of variance parameter}

\item{mc.cores}{number of cores for parallel}
}
\value{
A function of the probability parmater p
}
\description{
Empirical Bayes Meta-Analytic Prior for Binomial Data
}

