% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.MAP.FB.R
\name{binom.MAP.FB}
\alias{binom.MAP.FB}
\title{Full Bayesian Meta-Analytic Prior for Binomial Data}
\usage{
binom.MAP.FB(x, n, tau.prior, verbose = FALSE)
}
\arguments{
\item{x}{historical events}

\item{n}{historical trials}

\item{tau.prior}{optional prior on heterogeneity parameter}
}
\value{
A function of the probability parmater p
}
\description{
Full Bayesian Meta-Analytic Prior for Binomial Data
}
\examples{

}

