% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.EB.Beta.MC.R, R/binom.PP.EB.Beta.R
\name{binom.PP.EB.Beta}
\alias{binom.PP.EB.Beta}
\title{Empirical Bayes Power Prior for Binomial Data}
\usage{
binom.PP.EB.Beta(x, n, X, N, verbose = FALSE)

binom.PP.EB.Beta(x, n, X, N, verbose = FALSE)
}
\arguments{
\item{x}{historical events}

\item{n}{historical trials}

\item{x}{historical events}

\item{n}{historical trials}
}
\value{
A function of the probability parmater p

A function of the probability parmater p
}
\description{
Empirical Bayes Power Prior for Binomial Data

Empirical Bayes Power Prior for Binomial Data
}
\examples{


}

