% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.EB.R
\name{binom.PP.EB}
\alias{binom.PP.EB}
\title{Empirical Bayes Power Prior for Binomial Data}
\usage{
binom.PP.EB(x, n, X, N, verbose = FALSE, mc.cores = 1, p.prior.a = 1,
  p.prior.b = 1, max.dn = FALSE, mix = FALSE)
}
\arguments{
\item{x}{number of historical successes}

\item{n}{number historical patients}

\item{X}{vector of new successes to computer prior for}

\item{N}{number of new patients}

\item{verbose}{TRUE}

\item{mc.cores}{number of cores for parallel}

\item{p.prior.a}{shape1 parameter of initial beta prior for successes}

\item{p.prior.b}{shape2 parameter of initial beta prior for successes}

\item{max.dn}{Set a limit on the sum of the weights*n}

\item{mix}{if TRUE return a mixture.prior object otherwise the function}
}
\value{
Either a list of mixture.prior objects which can be evaluated with eval.mixture.prior or a density function of the probability parmater p given X.
}
\description{
Empirical Bayes Power Prior for Binomial Data
}
\examples{
\donttest{
xh <- c(30,40,50)
nh <- c(90,95,110)
eb <- binom.PP.EB(xh, nh, X=34,N=75, mix=TRUE) # EB power prior for new data 34/75.
}
}
