% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.EB.R
\name{binom.PP.EB}
\alias{binom.PP.EB}
\title{Empirical Bayes Power Prior for Binomial Data}
\usage{
binom.PP.EB(x, n, X, N, verbose = FALSE, mc.cores = 1, p.prior.a = 1,
  p.prior.b = 1)
}
\arguments{
\item{x}{number of historical successes}

\item{n}{number historical patients}

\item{X}{vector of new successes to computer prior for}

\item{N}{number of new patients}

\item{verbose}{TRUE}

\item{mc.cores}{number of cores for parallel}

\item{p.prior.a}{shape1 parameter of initial beta prior for successes}

\item{p.prior.b}{shape2 parameter of initial beta prior for successes}
}
\value{
A function of the probability parmater p
}
\description{
Empirical Bayes Power Prior for Binomial Data
}

