% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.EB.R
\name{binom.PP.EB}
\alias{binom.PP.EB}
\title{Empirical Bayes Power Prior for Binomial Data}
\usage{
binom.PP.EB(x, n, X, N, verbose = FALSE, mc.cores = 1)
}
\arguments{
\item{x}{historical events}

\item{n}{historical trials}

\item{tau.prior}{optional prior on heterogeneity parameter}
}
\value{
A function of the probability parmater p
}
\description{
Empirical Bayes Power Prior for Binomial Data
}
\examples{

}

