% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.EB.Sep.Mix.R
\name{binom.PP.EB.Sep}
\alias{binom.PP.EB.Sep}
\title{Empirical Bayes Power Prior for Binomial Data with the power parameters esimated separately}
\usage{
binom.PP.EB.Sep(x, n, X, N, verbose = FALSE, p.prior.a = 1, p.prior.b = 1,
  mix = FALSE)
}
\arguments{
\item{x}{number of historical successes}

\item{n}{number historical patients}

\item{X}{number of new successes}

\item{N}{number of new patients}

\item{verbose}{Print messages}

\item{p.prior.a}{shape1 parameter of initial beta prior for successes}

\item{p.prior.b}{shape2 parameter of initial beta prior for successes}

\item{mix}{if TRUE return a mixture.prior object otherwise the function}
}
\value{
Either a list of mixture.prior objects which can be evaluated with eval.mixture.prior or a density function of the probability parmater p given X.
}
\description{
Empirical Bayes Power Prior for Binomial Data with the power parameters esimated separately
}
\examples{
\donttest{
xh <- c(30,40,50)
nh <- c(90,95,110)
eb.sep <- binom.PP.EB.Sep(xh, nh, X=34,N=75, mix=TRUE)
}

}
