% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.EB.Sep.R
\name{binom.PP.EB.Sep}
\alias{binom.PP.EB.Sep}
\title{Empirical Bayes Power Prior for Binomial Data Esimated Separately}
\usage{
binom.PP.EB.Sep(x, n, X, N, verbose = FALSE)
}
\arguments{
\item{x}{historical events}

\item{n}{historical trials}

\item{X}{current events}

\item{N}{current number of trials}

\item{tau.prior}{optional prior on heterogeneity parameter}
}
\value{
A function of the probability parmater p
}
\description{
Empirical Bayes Power Prior for Binomial Data Esimated Separately
}
\examples{

}

