% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.FB.Cor.R
\name{binom.PP.FB.COR}
\alias{binom.PP.FB.COR}
\title{Full Bayes Power Prior with Correlated Weight Parameters using Rao-Blackwell}
\usage{
binom.PP.FB.COR(x, n, verbose = FALSE, mixture.size = 1000,
  d.prior.cor = 0, p.prior.a = 1, p.prior.b = 1, mc.cores = 1,
  mix = FALSE)
}
\arguments{
\item{x}{number of historical successes}

\item{n}{number historical patients}

\item{verbose}{Print messages}

\item{mixture.size}{Number of components in the mixture.}

\item{d.prior.cor}{Correlation parameter for transformed multivariate normal prior on weights}

\item{p.prior.a}{shape1 parameter for initial beta prior on probability}

\item{p.prior.b}{shape2 parameter for initial beta prior on probability}

\item{mc.cores}{Number of cores for parallel}

\item{mix}{return the mixture object}
}
\value{
A density function
}
\description{
Full Bayes Power Prior with Correlated Weight Parameters using Rao-Blackwell
}
\examples{
\donttest{
xh <- c(30,40,50)
nh <- c(90,95,110)

#fit a full Bayes power prior with 500 samples
fb <- binom.PP.FB.COR(x=xh, n=nh, mixture.size=500, mix=TRUE) 
mean.mixture.prior(fb) #calculate the mean
}


}
