% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.FB.MC.BE.R
\name{binom.PP.FB.MC.BE}
\alias{binom.PP.FB.MC.BE}
\title{Title}
\usage{
binom.PP.FB.MC.BE(x, n, verbose = FALSE, length = 30, d.prior.a = 1,
  d.prior.b = 1, mc.cores = 1, samples = 10000, focus)
}
\arguments{
\item{d.prior.a}{shape1 parameter for beta prior on weight}

\item{d.prior.b}{shape2 parameter for beta prior on weight}

\item{focus}{}
}
\description{
Title
}

