% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.FB.MC.COR.R
\name{binom.PP.FB.MC.COR}
\alias{binom.PP.FB.MC.COR}
\title{Title}
\usage{
binom.PP.FB.MC.COR(x, n, verbose = FALSE, length = 200, d.prior.cor = 0,
  p.prior.a = 1, p.prior.b = 1, mc.cores = 1, samples = 10000, focus)
}
\arguments{
\item{p.prior.a}{shape1 parameter for beta prior on probability}

\item{p.prior.b}{shape2 parameter for beta prior on probability}

\item{focus}{}

\item{d.prior.a}{shape1 parameter for beta prior on weight}

\item{d.prior.b}{shape2 parameter for beta prior on weight}
}
\description{
Title
}

