% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.FB.R
\name{binom.PP.FB}
\alias{binom.PP.FB}
\title{Full Bayes Power Prior for Binomial Data}
\usage{
binom.PP.FB(x, n, verbose = FALSE, length = 30, dprior, mc.cores = 1,
  p.prior.a = 1, p.prior.b = 1)
}
\arguments{
\item{x}{number of historical successes}

\item{n}{number historical patients}

\item{verbose}{Print messages}

\item{length}{Number of points to evaluate density at}

\item{dprior}{Density function for prior on weights}

\item{mc.cores}{Number of cores for parallel}

\item{p.prior.a}{shape1 parameter for beta prior on probability}

\item{p.prior.b}{shape2 parameter for beta prior on probability}
}
\value{
A function of the probability parmater p
}
\description{
Full Bayes Power Prior for Binomial Data
}

