% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.PP.FIX.R
\name{binom.PP.FIX}
\alias{binom.PP.FIX}
\title{Fixed Weight Power Prior for Binomial Data}
\usage{
binom.PP.FIX(x, n, d, p.prior.a = 1, p.prior.b = 1, verbose)
}
\arguments{
\item{x}{historical events}

\item{n}{historical trials}

\item{d}{fixed weight}

\item{p.prior.a}{shape1 parameter for beta prior on p}

\item{p.prior.b}{shape2 parameter for beta prior on p}

\item{verbose}{Print messages}
}
\value{
Probability density function for paramater p
}
\description{
Fixed Weight Power Prior for Binomial Data
}

