% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.prior.R
\name{binom.prior}
\alias{binom.prior}
\title{Prior distribution for binomial data}
\usage{
binom.prior(type = c("MAP.FB", "MAP.EB", "PP.FB", "PP.EB", "PP.EB.Beta",
  "Beta.EB", "PP.EB.Sep", "PP.Fix", "PP.Cor"), x, n, verbose = FALSE, ...)
}
\arguments{
\item{type}{Which prior to provide}

\item{x}{Vector of historical events}

\item{n}{Vector of historical sample sizes}

\item{...}{Parameters to provide to prior}
}
\value{
A function
}
\description{
Prior distribution for binomial data
}

