% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coverage.R
\name{calc.coverage}
\alias{calc.coverage}
\title{Coverage using prior using the Bayarri thing}
\usage{
calc.coverage(prior, level, n.control, smooth, ...)
}
\arguments{
\item{length}{}
}
\description{
Coverage using prior using the Bayarri thing
}

