% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.mean.R
\name{calc.MSE.mean}
\alias{calc.MSE.mean}
\title{Calculate mean squared error based on the mean of the posterior}
\usage{
calc.MSE.mean(prior, prob.range = c(0.5, 1), length = 20, n.binom = 30,
  mc.cores = 1, posterior)
}
\arguments{
\item{n.binom}{}
}
\description{
Calculate mean squared error based on the mean of the posterior
}

