% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.mode.R
\name{calc.MSE.mode}
\alias{calc.MSE.mode}
\title{Calculate mean squared error based on the mode of the posterior}
\usage{
calc.MSE.mode(prior, prob.range = c(0.5, 1), length = 20, n.binom = 30,
  mc.cores = 1)
}
\arguments{
\item{n.binom}{}
}
\description{
Calculate mean squared error based on the mode of the posterior
}

