% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE1.R
\name{calc.MSE}
\alias{calc.MSE}
\title{Calculate mean squared error}
\usage{
calc.MSE(prior, prob.range = c(0.5, 1), length = 20, n.binom = 30,
  mc.cores = 1)
}
\arguments{
\item{n.binom}{}
}
\description{
Calculate mean squared error
}

