% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{calc.power}
\alias{calc.power}
\title{Calculate power}
\usage{
calc.power(
  prior,
  prob.range = c(0.5, 0.9),
  length = 20,
  n.binom.control = 30,
  n.binom.treatment = n.binom.control,
  treatment.difference = 0.1,
  level = 0.95,
  sig.mat,
  LR = FALSE
)
}
\arguments{
\item{prior}{Prior to calculate posterior}

\item{prob.range}{Range of values to calculate over}

\item{length}{Number of values to calculate for}

\item{n.binom.control}{Number of patients in new trial's contral arm}

\item{n.binom.treatment}{Number of patients in new trial's treatment arm}

\item{treatment.difference}{Predefined treatment difference}

\item{level}{Significance level for calculation of sig.mat}

\item{sig.mat}{Precalculated significance matrix (See sig.matrix)}

\item{LR}{Calculate the likelihood ratio of power/type I error}
}
\value{
Vector or power values
}
\description{
Calculate power
}
\examples{
\donttest{
xh <- c(30,40,50)
nh <- c(90,95,110)
fix <- binom.PP.FIX(x=xh, n=nh, d=c(.2,.5,.7), mix=TRUE) # set different weights
mat <- sig.matrix(n.control=50, n.treatment=75, prior = fix)

# Calculate the power to detect difference of 0.15
pow <- calc.power(sig.mat=mat, prob.range=c(.5,.85),
  n.binom.control = 50, n.binom.treatment=75, treatment.difference=0.15)

# Calculate the type 1 error by setting treatment.difference = 0
pow <- calc.power(sig.mat=mat, prob.range=c(.5,.85),
  n.binom.control = 50, n.binom.treatment=75, treatment.difference=0)

}
}
