% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{calc.power}
\alias{calc.power}
\title{Calculate power}
\usage{
calc.power(prior, prob.range = c(0.5, 0.9), length = 20,
  n.binom.control = 30, n.binom.treatment = n.binom.control,
  treatment.difference = 0.1, level = 0.95, sig.mat, mc.cores = 1)
}
\arguments{
\item{n.binom}{}
}
\description{
Calculate power
}

