% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.3.R
\name{conj.approx}
\alias{conj.approx}
\title{Approximate distribution with mixture model}
\usage{
conj.approx(distr, type = c("beta", "normal"), max.degree = 3,
  min.degree = max.degree, robust = FALSE, range = c(0, 1),
  do.plot = FALSE, return.value = 0.2, starts, length.fit = 100)
}
\arguments{
\item{distr}{The density function of the function to be approximated}

\item{type}{Use beta or normal densities in the mixture}

\item{max.degree}{Maximum size of mixture}

\item{min.degree}{Minimum size of mixture}

\item{robust}{Include a robust component? (FALSE or final weight between 0 and 1)}

\item{range}{The range of the density function to approximate}

\item{do.plot}{Show the components of the mixture in a plot}

\item{return.value}{Convergence threshold}

\item{starts}{Initial parameters for the search}

\item{length.fit}{Number of points to fit to}
}
\value{
Returns a mixture.prior object
}
\description{
Approximate distribution with mixture model
}

