% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.R
\name{conj.approx}
\alias{conj.approx}
\title{Mixture distribution}
\usage{
conj.approx(distr, type = c("beta", "normal"), degree = 3, range = c(0,
  1), starts)
}
\arguments{
\item{range}{}
}
\description{
Mixture distribution
}

