% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{ess.mixture.prior}
\alias{ess.mixture.prior}
\title{Calculate the equivalent sample size of mixture model}
\usage{
ess.mixture.prior(mixture.prior)
}
\arguments{
\item{mixture.prior}{Mixture prior object}
}
\value{
Sample size
}
\description{
Calculate the equivalent sample size of mixture model
}

