% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.rw.R
\name{mean.mixture.prior}
\alias{mean.mixture.prior}
\title{Calculate the mean of mixture model}
\usage{
\method{mean}{mixture.prior}(x, ...)
}
\arguments{
\item{x}{Mixture prior object}

\item{...}{Ignored}
}
\value{
Mean of prior
}
\description{
Calculate the mean of mixture model
}
\examples{
\donttest{
xh <- c(30,40,50)
nh <- c(90,95,110)

 #fit a full Bayes power prior with 500 samples
fb <- binom.PP.FB.COR(x=xh, n=nh, mixture.size=500, mix=TRUE)
mean.mixture.prior(fb) #calculate the mean
}
}
