% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.PP.FB.INLA.R
\name{normal.PP.FB.INLA}
\alias{normal.PP.FB.INLA}
\title{Normal power prior with full Bayes by INLA}
\usage{
normal.PP.FB.INLA(x, sd, verbose = FALSE)
}
\arguments{
\item{x}{numeric vector of results}

\item{sd}{standard deviations}

\item{verbose}{Print messages}
}
\value{
A density function
}
\description{
Normal power prior with full Bayes by INLA
}

