% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.PP.FB.MC.R
\name{normal.PP.FB.MC}
\alias{normal.PP.FB.MC}
\title{Normal power prior with full Bayes by Monte Carlo}
\usage{
normal.PP.FB.MC(x, sd, verbose = FALSE, length = 30, d.prior.a = 1,
  d.prior.b = 1, mc.cores = 1, samples = 10000, focus)
}
\arguments{
\item{x}{numeric vector of results}

\item{sd}{standard deviations}

\item{focus}{list(from, to, length) to evaluate the marginal density when fitting the curve}
}
\description{
Normal power prior with full Bayes by Monte Carlo
}

