% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.rw.R
\name{plot.mixture.prior}
\alias{plot.mixture.prior}
\title{Plot mixture model}
\usage{
\method{plot}{mixture.prior}(x, ..., at = seq(0, 1, 0.01), stack = FALSE,
  lines.only = FALSE)
}
\arguments{
\item{x}{a mixture prior object to plot}

\item{...}{Additional arguments to \code{\link{lines}}}

\item{at}{Vector of values to evaluate density at}

\item{stack}{Plot the mixture components as stacked regions}

\item{lines.only}{Draw lines only (to be used over an existing plot)}
}
\description{
Plot mixture model
}
