% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{plot.mixture.prior}
\alias{plot.mixture.prior}
\title{Plot mixture model}
\usage{
\method{plot}{mixture.prior}(mixture.prior, x, stack = FALSE,
  lines.only = FALSE, ...)
}
\arguments{
\item{mixture.prior}{Mixture prior}

\item{x}{Vector of values to evaluate density at}

\item{stack}{Plot the mixture components as stacked regions}

\item{lines.only}{Draw lines only (to be used over an existing plot)}

\item{...}{Additional arguments to \code{\link{lines}}}
}
\description{
Plot mixture model
}

