% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{posterior.mixture.prior}
\alias{posterior.mixture.prior}
\title{Calculate posterior mixture model}
\usage{
posterior.mixture.prior(xs, ns, mixture.prior)
}
\arguments{
\item{xs}{Number of successes in new trial}

\item{ns}{Number of patients in new trial}

\item{mixture.prior}{Mixture prior object}
}
\value{
A \code{mixture.prior} object
}
\description{
Calculate posterior mixture model
}

