% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.R
\name{update.mixture.prior}
\alias{update.mixture.prior}
\title{Update parameters or weights in mixture model}
\usage{
\method{update}{mixture.prior}(object, ...)
}
\arguments{
\item{object}{mixture.prior to update}

\item{...}{Additional arguments: pars and weights. See \code{\link{create.mixture.prior}}}
}
\value{
A \code{mixture.prior} object with updated weights and parameters
}
\description{
Update parameters or weights in mixture model
}

