% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.list.rw.R
\name{var.mixture.prior}
\alias{var.mixture.prior}
\title{Calculate the variance of mixture model}
\usage{
var.mixture.prior(mixture.prior)
}
\arguments{
\item{mixture.prior}{Mixture prior object}
}
\value{
Variance of prior
}
\description{
Calculate the variance of mixture model
}
\examples{
\donttest{
xh <- c(30,40,50)
nh <- c(90,95,110)

 #fit a full Bayes power prior with 500 samples
fb <- binom.PP.FB(x=xh, n=nh, mixture.size=500, mix=TRUE)

 #fit a full Bayes power prior with 500 samples with correlation
fb2 <- binom.PP.FB(x=xh, n=nh, mixture.size=500, mix=TRUE,d.prior.cor=0.9)

# calculate the variances of the distributions
var.mixture.prior(fb2)
var.mixture.prior(fb)
}

}
